- The javascript properly executes now.
Important:

When dealing with chrome extensions, we have to add listeners to the javascript files themselves as they share the DOM of the page that they are associated with. So, instead of calling startTime() inline in index.html, I moved it to production.js as an event listener to elements that have the .background class, which in this case would be the body tag. I would refer to the last comment in the following StackOverflow question in order to get a better idea of what I'm talking about - http://stackoverflow.com/questions/17933234/chrome-extension-how-to-a-call-a-function-defined-in-content-script.

Other Changes:

Temporarily removed the icon from the manifest file since it wasn't allowing me to load the extension