- - Main Addition - -
New bookmarks will be saved to Google Chrome when the TEST button is pressed. When you leave the page and open a new tab page, those new bookmarks will be accessible and will be updated if you add any additional bookmarks (be sure to press the test button again)

- - Other Changes - -

For testing purposes, I recreated the index with Ashish's bookmark code.

Modified the manifest so that I could make use of Chrome's storage API (addition of the storage permission)